# Add option to ignore reference cycles on serialization