--- a/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/ReferenceHandlingStrategy.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/ReferenceHandlingStrategy.cs
@@ -7,6 +7,6 @@ namespace System.Text.Json.Serialization
     {
         None,
         Preserve,
-        IgnoreCycle
+        IgnoreCycles
     }
 }