… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonPropertyInfoOfT.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Read.HandlePropertyName.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Write.HandleMetadata.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializerOptions.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/ReadStack.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/ReferenceEqualsWrapper.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/ReferenceHandler.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/ReferenceResolver.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Serialization/WriteStack.cs
… src/libraries/System.Text.Json/tests/Serialization/ReferenceHandlerTests.IgnoreCycle.cs
… src/libraries/System.Text.Json/tests/Serialization/ReferenceHandlerTests.IgnoreCycles.cs
… src/libraries/System.Text.Json/tests/System.Text.Json.Tests.csproj