Fixes https://github.com/dotnet/runtime/issues/40099

This is a draft implementing a new API for `ReferenceHandler` capable of detect circularity on the input value of JsonSerializer.Serialize. A sample usage scenario would be the following:

```cs
class Node
{
    public string Description { get; set; }
    public object Next { get; set; }
}

void Test()
{
    var node = new Node { Description = "Node 1" };
    node.Next = node;
    
    var opts = new JsonSerializerOptions { ReferenceHandler = ReferenceHandler.IgnoreCycle };
    
    string json = JsonSerializer.Serialize(node, opts);
    Console.WriteLine(json); // Prints {"Description":"Node 1","Next":null}
}
```

**Note that reference cycles are broken by null tokens in JSON.**