@eiriktsarpalis 2020-12-16
I think this has already been discussed in the parent issue, but as a json user, assuming I didn't want to preserve references I would prefer it if the serializer threw an exception rather than silently ignoring data (because there's likely a bug in my code).