@Jozkee 2020-12-16
@eiriktsarpalis this is not changing the default behavior that throws an exception when MaxDepth is reached. You have to opt-in for ignoring reference loops as shown in the description's code snippet.