@steveharter 2021-01-12
@Jozkee have you evaluated what it would take to improve the existing extensibility model to support ignore? That would either be an alternative to this feature or perhaps complementary to handle scenarios with different requirements.