@Jozkee 2021-01-13
@steveharter I was thinking that we could add a ReferenceCycleStrategy flag that indicates the semantics that will be used and add a couple methods to ReferenceResolver for such purpose.

```diff
public abstract class ReferenceResolver
{
    protected ReferenceResolver() { }
    public abstract void AddReference(string referenceId, object value);
    public abstract string GetReference(object value, out bool alreadyExists);
    public abstract object ResolveReference(string referenceId);

+   public virtual bool PushReferenceForCycleDetection(object value);
+   public virtual void PopReferenceForCycleDetection(object value);
}

public abstract class ReferenceHandler
{
    public static ReferenceHandler Preserve { get { throw null; } }
    public static ReferenceHandler IgnoreCycle { get { throw null; } }
    public abstract ReferenceResolver CreateResolver();

+   public virtual JsonReferenceCycleStrategy ReferenceCycleStrategy => JsonReferenceCycleStrategy.None;
}

+public enum JsonReferenceCycleStrategy
+{
+   None, // Use preserve semantics.
+   Ignore // Use ignore semantics.
+}
```