@Jozkee 2021-01-21
```diff
-public abstract class ReferenceResolver
+public abstract class ReferenceResolver : ReferenceResolverBase
{
    public abstract void AddReference(string referenceId, object value);
    public abstract string GetReference(object value, out bool alreadyExists);
    public abstract object ResolveReference(string referenceId);
}

+public abstract class ReferenceResolverBase {}

+public abstract class IgnoreReferenceResolver : ReferenceResolverBase 
+{
+   public abstract void PopReferenceForCycleDetection() => throw new InvalidOperationException();
+   public abstract void PushReferenceForCycleDetection(object value) => throw new InvalidOperationException();
+   public abstract bool ContainsReferenceForCycleDetection(object value) => throw new InvalidOperationException();
+}

// Binary breaking change.
-public sealed class ReferenceHandler<T> : ReferenceHandler where T : ReferenceResolver, new()
+public sealed class ReferenceHandler<T> : ReferenceHandler where T : ReferenceResolverBase, new()
{
    
-   public override ReferenceResolver CreateResolver() { throw null; }
+   public override ReferenceResolverBase CreateResolver() { throw null; }
}

public abstract partial class ReferenceHandler
{
    protected ReferenceHandler() { }
    public static ReferenceHandler Preserve { get; }
    public static ReferenceHandler IgnoreCycle { get; }
-   public abstract ReferenceResolver CreateResolver();
+   public abstract ReferenceResolverBase CreateResolver();
}
```
I was thinking a bit more about how an uber class could help us to keep single responsibility on each resolver and thought that we could probably add `ReferenceResolverBase` and make `ReferenceResolver` inherit from it, that way, we can keep it to work as a set of references that uses the Preserve semantics.

`ReferenceResolverBase` would allow us to add a new resolver (`IgnoreReferenceResolver`) that can be used as a stack of references and would use the ignore semantics from Newtonsoft.

However the downside of such design is that it would introduce a binary breaking change on users that already implement their own `ReferenceHandler`. 

@steveharter  @ahsonkhan @layomia @eiriktsarpalis thoughts?