@Jozkee 2021-01-21
Update: addressed feedback and changed feature behavior to treat reference loops as null values in order to not alter the length of collections.