@Jozkee 2021-01-26
Fixed a perf regression caused by using `IsIgnoreCyclesEnabled` method in order to constantly check if the feature was on.
https://github.com/dotnet/runtime/blob/2e55d738d9899d9623c809f8d79479dc081fa524/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Write.HandleMetadata.cs#L91-L92

New internal field `ReferenceHandlingStrategy` in `JsonSerializerOptions` is used instead in order to perform the same check. Using a field that refreshes only when `ReferenceHandler` changes is cheaper than using above method.
https://github.com/dotnet/runtime/blob/1905462b9b48b694eb898941d4f1019c14dbdce5/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializerOptions.cs#L491-L496