@Jozkee 2021-01-26
``` ini

BenchmarkDotNet=v0.12.1.1466-nightly, OS=Windows 10.0.18363.1316 (1909/November2019Update/19H2)
Intel Core i7-9750H CPU 2.60GHz, 1 CPU, 12 logical and 6 physical cores
.NET SDK=5.0.100
  [Host]     : .NET 5.0.0 (5.0.20.51904), X64 RyuJIT
  Job-ABOOKC : .NET 6.0.0 (42.42.42.42424), X64 RyuJIT
  Job-QJROBB : .NET 6.0.0 (42.42.42.42424), X64 RyuJIT

PowerPlanMode=00000000-0000-0000-0000-000000000000  Arguments=/p:DebugType=portable  IterationTime=250.0000 ms  
MaxIterationCount=20  MinIterationCount=15  WarmupCount=1  

```
|                                  Type |               Method |        Job |                                                                                                  Toolchain |         Mean |       Error |      StdDev |       Median |          Min |          Max | Ratio | MannWhitney(3ms) |   Gen 0 |  Gen 1 | Gen 2 | Allocated |
|-------------------------------------- |--------------------- |----------- |----------------------------------------------------------------------------------------------------------- |-------------:|------------:|------------:|-------------:|-------------:|-------------:|------:|----------------- |--------:|-------:|------:|----------:|
|             WriteJson&lt;IndexViewModel&gt; | SerializeToUtf8Bytes | Job-ABOOKC | \runtime2\artifacts\bin\testhost\net6.0-windows-Release-x64\shared\Microsoft.NETCore.App\6.0.0\CoreRun.exe |  17,076.1 ns |   172.47 ns |   161.33 ns |  17,057.3 ns |  16,803.3 ns |  17,401.4 ns |  1.01 |             Same |  2.0292 | 0.0676 |     - |  13,024 B |
|             WriteJson&lt;IndexViewModel&gt; | SerializeToUtf8Bytes | Job-QJROBB |  \runtime\artifacts\bin\testhost\net6.0-windows-Release-x64\shared\Microsoft.NETCore.App\6.0.0\CoreRun.exe |  16,887.4 ns |    99.72 ns |    93.27 ns |  16,875.5 ns |  16,746.1 ns |  17,076.6 ns |  1.00 |             Base |  2.0380 | 0.0679 |     - |  13,024 B |
|                                       |                      |            |                                                                                                            |              |             |             |              |              |              |       |                  |         |        |       |           |
|                      WriteJson&lt;Int32&gt; | SerializeToUtf8Bytes | Job-ABOOKC | \runtime2\artifacts\bin\testhost\net6.0-windows-Release-x64\shared\Microsoft.NETCore.App\6.0.0\CoreRun.exe |     122.9 ns |     1.55 ns |     1.37 ns |     123.1 ns |     119.2 ns |     125.0 ns |  1.01 |             Same |  0.0289 |      - |     - |     184 B |
|                      WriteJson&lt;Int32&gt; | SerializeToUtf8Bytes | Job-QJROBB |  \runtime\artifacts\bin\testhost\net6.0-windows-Release-x64\shared\Microsoft.NETCore.App\6.0.0\CoreRun.exe |     121.1 ns |     1.78 ns |     1.48 ns |     121.1 ns |     117.9 ns |     123.5 ns |  1.00 |             Base |  0.0289 |      - |     - |     184 B |
|                                       |                      |            |                                                                                                            |              |             |             |              |              |              |       |                  |         |        |       |           |
|  WriteJson&lt;LargeStructWithProperties&gt; | SerializeToUtf8Bytes | Job-ABOOKC | \runtime2\artifacts\bin\testhost\net6.0-windows-Release-x64\shared\Microsoft.NETCore.App\6.0.0\CoreRun.exe |     627.5 ns |     4.92 ns |     4.60 ns |     627.4 ns |     621.2 ns |     636.3 ns |  1.01 |             Same |  0.0578 |      - |     - |     376 B |
|  WriteJson&lt;LargeStructWithProperties&gt; | SerializeToUtf8Bytes | Job-QJROBB |  \runtime\artifacts\bin\testhost\net6.0-windows-Release-x64\shared\Microsoft.NETCore.App\6.0.0\CoreRun.exe |     624.2 ns |     4.57 ns |     4.28 ns |     624.6 ns |     617.5 ns |     632.7 ns |  1.00 |             Base |  0.0581 |      - |     - |     376 B |
|                                       |                      |            |                                                                                                            |              |             |             |              |              |              |       |                  |         |        |       |           |
|                   WriteJson&lt;Location&gt; | SerializeToUtf8Bytes | Job-ABOOKC | \runtime2\artifacts\bin\testhost\net6.0-windows-Release-x64\shared\Microsoft.NETCore.App\6.0.0\CoreRun.exe |     761.1 ns |     4.59 ns |     4.07 ns |     760.9 ns |     755.2 ns |     770.5 ns |  0.98 |             Same |  0.0611 |      - |     - |     384 B |
|                   WriteJson&lt;Location&gt; | SerializeToUtf8Bytes | Job-QJROBB |  \runtime\artifacts\bin\testhost\net6.0-windows-Release-x64\shared\Microsoft.NETCore.App\6.0.0\CoreRun.exe |     778.0 ns |     7.14 ns |     6.68 ns |     778.5 ns |     768.6 ns |     792.3 ns |  1.00 |             Base |  0.0596 |      - |     - |     384 B |
|                                       |                      |            |                                                                                                            |              |             |             |              |              |              |       |                  |         |        |       |           |
|             WriteJson&lt;LoginViewModel&gt; | SerializeToUtf8Bytes | Job-ABOOKC | \runtime2\artifacts\bin\testhost\net6.0-windows-Release-x64\shared\Microsoft.NETCore.App\6.0.0\CoreRun.exe |     314.8 ns |     2.71 ns |     2.54 ns |     314.4 ns |     310.9 ns |     320.0 ns |  0.98 |             Same |  0.0410 |      - |     - |     264 B |
|             WriteJson&lt;LoginViewModel&gt; | SerializeToUtf8Bytes | Job-QJROBB |  \runtime\artifacts\bin\testhost\net6.0-windows-Release-x64\shared\Microsoft.NETCore.App\6.0.0\CoreRun.exe |     320.0 ns |     1.60 ns |     1.42 ns |     320.2 ns |     317.1 ns |     323.1 ns |  1.00 |             Base |  0.0416 |      - |     - |     264 B |
|                                       |                      |            |                                                                                                            |              |             |             |              |              |              |       |                  |         |        |       |           |
|    WriteJson&lt;MyEventsListerViewModel&gt; | SerializeToUtf8Bytes | Job-ABOOKC | \runtime2\artifacts\bin\testhost\net6.0-windows-Release-x64\shared\Microsoft.NETCore.App\6.0.0\CoreRun.exe | 413,696.7 ns | 3,250.95 ns | 3,040.94 ns | 412,356.4 ns | 410,011.2 ns | 418,388.0 ns |  1.01 |             Same | 29.6053 | 4.9342 |     - | 191,665 B |
|    WriteJson&lt;MyEventsListerViewModel&gt; | SerializeToUtf8Bytes | Job-QJROBB |  \runtime\artifacts\bin\testhost\net6.0-windows-Release-x64\shared\Microsoft.NETCore.App\6.0.0\CoreRun.exe | 408,065.2 ns | 3,297.89 ns | 3,084.85 ns | 408,330.6 ns | 403,268.6 ns | 412,867.3 ns |  1.00 |             Base | 28.8462 | 4.8077 |     - | 191,705 B |
|                                       |                      |            |                                                                                                            |              |             |             |              |              |              |       |                  |         |        |       |           |
| WriteJson&lt;SimpleStructWithProperties&gt; | SerializeToUtf8Bytes | Job-ABOOKC | \runtime2\artifacts\bin\testhost\net6.0-windows-Release-x64\shared\Microsoft.NETCore.App\6.0.0\CoreRun.exe |     232.2 ns |     1.87 ns |     1.75 ns |     231.9 ns |     229.1 ns |     235.3 ns |  1.01 |             Same |  0.0361 |      - |     - |     232 B |
| WriteJson&lt;SimpleStructWithProperties&gt; | SerializeToUtf8Bytes | Job-QJROBB |  \runtime\artifacts\bin\testhost\net6.0-windows-Release-x64\shared\Microsoft.NETCore.App\6.0.0\CoreRun.exe |     230.6 ns |     2.59 ns |     2.42 ns |     230.4 ns |     227.7 ns |     236.1 ns |  1.00 |             Base |  0.0366 |      - |     - |     232 B |